# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elegant Sip API",
    "version": "0.1.0",
    "description": "Backend for the Elegant Sip storefront — single-origin Darjeeling tea.\n\n**Money is integer paise** (1 rupee = 100 paise). Razorpay settles in paise and\nfloats cannot represent money. Totals are always a whole number of rupees:\nGST is rounded with the same rule the storefront uses, so the figure a customer\nis shown and the figure they are charged cannot drift apart.\n\nErrors share one shape across every endpoint, modelled on RFC 9457. `detail` is\nalways safe to display; `requestId` ties a response to the server logs.",
    "contact": {
      "name": "Elegant Sip",
      "email": "elegantsipdarjeeling@gmail.com"
    }
  },
...
  "paths": {
...
    "/v1/products/{slug}/reviews": {
      "get": {
        "summary": "Published reviews for a product",
        "tags": [
          "Catalogue"
        ],
        "description": "Only moderated reviews are returned. `verified` means a paid order from that customer actually contained this product — it is derived server-side and can never be self-asserted.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            },
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "author": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 5
                          },
                          "text": {
                            "type": "string"
                          },
                          "publishedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "verified": {
                            "type": "boolean",
                            "description": "True only when a paid order from this customer contains the product"
                          }
                        },
                        "required": [
                          "id",
                          "author",
                          "rating",
                          "text",
                          "publishedAt",
                          "verified"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "average": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 5
                    },
                    "count": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "reviews",
                    "average",
                    "count"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Write a review",
        "tags": [
          "Account"
        ],
        "description": "Requires a signed-in account. **`verified` is decided here**, by checking whether a\npaid order from this customer actually contains the product — the client cannot assert it.\n\nReviews are held unpublished until moderated, so the product page never shows\nsomething nobody has read.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "text": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": [
                  "rating",
                  "text"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            },
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "review": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "author": {
                          "type": "string"
                        },
                        "rating": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5
                        },
                        "text": {
                          "type": "string"
                        },
                        "verified": {
                          "type": "boolean",
                          "description": "True only when a paid order from this customer contains the product"
                        },
                        "publishedAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "author",
                        "rating",
                        "text",
                        "verified",
                        "publishedAt"
                      ],
                      "additionalProperties": false
                    },
                    "pendingModeration": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "review",
                    "pendingModeration"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/wishlist": {
      "get": {
        "summary": "Your wishlist",
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "productSlugs": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80,
                        "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                        "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                      }
                    }
                  },
                  "required": [
                    "productSlugs"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace your wishlist",
        "tags": [
          "Account"
        ],
        "description": "Idempotent whole-list replace. Used to merge a guest wishlist from localStorage on first sign-in, then to keep it in sync.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productSlugs": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                    }
                  }
                },
                "required": [
                  "productSlugs"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "productSlugs": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80,
                        "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                        "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                      }
                    }
                  },
                  "required": [
                    "productSlugs"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/newsletter": {
      "post": {
        "summary": "Subscribe to the newsletter",
        "tags": [
          "Account"
        ],
        "description": "Returns the welcome code in the response. The storefront shows it on screen rather than promising it by email — which it always did, and which is now also true of the API.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "welcomeCode": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "welcomeCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/contact": {
      "post": {
        "summary": "Send a message",
        "tags": [
          "Account"
        ],
        "description": "Delivers to the shop inbox. `delivered` reports whether the message actually went out — the UI must not claim it was sent when this is false.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "subject": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "message": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "name",
                  "email",
                  "subject",
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "delivered": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "delivered"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.elegantsip.com",
      "description": "Production"
    }
  ],
...
}